# Optimized SD3 pipeline