--- a/examples/stable-diffusion/README.md
+++ b/examples/stable-diffusion/README.md
@@ -358,7 +358,7 @@ python text_to_image_generation.py \
     --use_hpu_graphs \
     --gaudi_config Habana/stable-diffusion \
     --sdp_on_bf16 \
-    --bf16
+    --bf16 \
     --quant_mode quantize
 ```