--- a/examples/stable-diffusion/quantization/stable-diffusion-3/measure_config.json
+++ b/examples/stable-diffusion/quantization/stable-diffusion-3/measure_config.json
@@ -1,0 +1,5 @@
+{
+    "method": "HOOKS",
+    "mode": "MEASURE",
+    "dump_stats_path": "quantization/stable-diffusion-3/measure_all/fp8"
+}