# What does this PR do?

 * HPU graphs enabled
 * Batching for inference enabled
 * Fused SDPA integrated
 * FP8 quantization enabled
 * Updated run command in ReadMe and added documentation to run FP8


## Performance comparison of Pre and Post Optimizations for SD3 Pipeline
**Achieved ~4x throughput improvement with HPU Graph and Fused SDPA**

| **Batch Size** | **No of Images** | **Batches** | **Pre-Optimization Throughput (samples/sec)** | **Optimized Throughput (samples/sec)** |
|:--------------:|:----------------:|:-----------:|:--------------------------------------------:|:---------------------------------------:|
|       1        |        10        |      10     |                   0.047                      |                0.227                   |
|       2        |        20        |      10     |                   0.051                      |                0.228                   |
|       4        |        40        |      10     |                   0.052                      |                0.228                   |
|       8        |        80        |      10     |                   0.052                      |                0.227                   |



## Diffusers CI Tests Pass

```
================================================================================= test session starts ==================================================================================
platform linux -- Python 3.10.12, pytest-7.4.4, pluggy-1.5.0
rootdir: /mnt/deepak_oh_new
configfile: setup.cfg
collected 165 items                                                                                                                                                                    

tests/test_diffusers.py .......sssss..........s........s.....s..................s.ss.sssssssssssss.............sss.......s....s.....ss.s.ssssss......s...s......s.sss........s.. [ 92%]
s.s....s.s...        
============================================================== 116 passed, 49 skipped, 280 warnings in 1495.78s (0:24:55) ==============================================================

```


This PR is jointly co-authored with:
- Daniel Socek <daniel.socek@intel.com>

## Before submitting
- [ ] This PR fixes a typo or improves the docs (you can dismiss the other checks if that's the case).
- [x] Did you make sure to update the documentation with your changes?
- [ ] Did you write any new necessary tests?