@dsocek 2025-01-08
Additional performance results of this PR:

Device | Docker | Mode | Num Inference Steps | Batch Size | Sec/Batch Step | Steps/sec | Samples/sec | Time for 1 Image
-- | -- | -- | -- | -- | -- | -- | -- | --
Gaudi3 | 1.19.0 | Lazy w/ HPU Graphs | 28 | 1 | 0.09 | 10.64 | 0.380 | 2.63
Gaudi2 | 1.19.0 | Lazy w/ HPU Graphs | 28 | 1 | 0.15 | 6.48 | 0.231 | 4.32
Gaudi2 | 1.19.0 | Lazy w/ HPU Graphs | 28 | 10 | 1.53 | 6.54 | 0.234 | 4.28