@sywangyi 2025-01-10
[error][tid:C62] FP32 operations are not supported on this device. Node Name BatchGemm135