@dsocek 2025-01-10
> @deepak-gowda-narayana I see in the PR with torch.autocast(device_type="hpu", dtype=torch.bfloat16, enabled=self.gaudi_config.use_torch_autocast): this logic is removed, it will lead to crash if "--bf16" is not in commandline in Gaudi2D platform which only contains BF16 gemm. error like [INFO|pipeline_stable_diffusion_3.py:569] 2025-01-10 05:26:37,344 >> 1 prompt(s) received, 10 generation(s) per prompt, 1 sample(s) per batch, 10 total batch(es). Traceback (most recent call last): File "/root/optimum-habana/examples/stable-diffusion/text_to_image_generation.py", line 701, in main() File "/root/optimum-habana/examples/stable-diffusion/text_to_image_generation.py", line 664, in main outputs = pipeline(prompt=args.prompts, **kwargs_call) File "/usr/local/lib/python3.10/dist-packages/torch/utils/_contextlib.py", line 116, in decorate_context return func(*args, **kwargs) File "/root/optimum-habana/optimum/habana/diffusers/pipelines/stable_diffusion_3/pipeline_stable_diffusion_3.py", line 610, in **call** ht.hpu.synchronize() File "/usr/local/lib/python3.10/dist-packages/habana_frameworks/torch/hpu/**init**.py", line 154, in synchronize return _hpu_C.synchronize_device() RuntimeError: synNodeCreateWithId failed for node: batch_gemm with synStatus 26 [Generic failure]. .
> 
> @heyuanliu-intel

@sywangyi Thank you for pointing this out.   We removed the `with torch.autocast(device_type="hpu", dtype=torch.bfloat16, enabled=self.gaudi_config.use_torch_autocast):` to allow users to use autocasting (and in addition we saw slight performance boost).  Removing `--bf16` works without issues on regular G2/G3.  Rather than putting a blanket BF16 restriction on all ops, maybe better strategy should be to put a note saying that on some more limited Gaudi devices (e.g. Gaudi2D) please use `--bf16` as some ops do not support FP32.  This way we still allow users to do more fine grained optimizations with custom ops lists (some ops on BF16 and some of FP32).